Avoid confusion around total vs pre-node time
This benchmark has 1000 nodes, takes 1.56 microseconds and thus the
graph shows 1.56 nanoseconds. Showing the per-node timings is
generally good (since that's most relevant here), but the purpose of
this paragraph was to also not forget about the total duration.

It can however be a bit confusing to see the exact "1.56 microseconds"
and thus believe it's a typo of the number in the graph.

Fixes #1